Finished game loop
Finished the game loop
Added Update / Render methods in Engine class
Added Init methods to both Game class and Engine class

Note: at line 51 in the Engine class is the run method -game loop-. I
normally initialize the beforeTime and currentTime variables to 0 and it
works perfectly in the logic for checking when a single second has
passed within the loop. Now, for about 2 seconds, the loop's second
logic is really bad, but corrects itself. I am really confused as to why
it is doing this when in previously I have never had this problem.